kafka-consume-produce/requirements.txt
kafka-python==2.0.1
packaging==20.4
protobuf==3.12.4
protobuf3==0.2.1
pyparsing==2.4.7
six==1.15.0
